% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_description}
\alias{get_description}
\title{Utility tools to get specimen description and images from NYBG's virtual herbarium}
\usage{
get_description(ny_id)
}
\arguments{
\item{ny_id}{The NYBG specimen ID. It is equivalent to the field catalogNumber
in GBIF.}

\item{path}{Where to save downloaded images. Defaults to getwd().}

\item{...}{Other parameters passed on to rgbif::occ_search.}
}
\description{
Utility tools to get specimen description and images from NYBG's virtual herbarium
}
\examples{
\dontrun{
get_description(1889108)
specimen_page(1889108)
fill_description(1889108)
get_image_url(1889108)
open_image(1889108)
download_image(1889108)
}
}

